CF-2 Renamed bucket package
