+++ src/main/java/com/bsjhx/cashflow/adapters/inbound/rest/BucketController.java
@@ -9,7 +9,7 @@ import com.bsjhx.cashflow.application.bucket.command.MoneyTransferCommandHandler
 import com.bsjhx.cashflow.application.bucket.command.OpenBucketCommandHandler;
 import com.bsjhx.cashflow.application.bucket.query.CurrentBalanceQuery;
 import com.bsjhx.cashflow.application.bucket.query.CurrentBalanceQueryHandler;
-import com.bsjhx.cashflow.domain.bucket.Money;
+import com.bsjhx.cashflow.domain.tracksheet.Money;
 import com.bsjhx.cashflow.domain.common.Event;
 import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.*;
