+++ src/main/java/com/bsjhx/cashflow/domain/tracksheet/Money.java
@@ -1,9 +1,7 @@
-package com.bsjhx.cashflow.domain.bucket;
+package com.bsjhx.cashflow.domain.tracksheet;
 
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
-
-import java.util.Objects;
 
 @Getter
 @EqualsAndHashCode
